requirements.txt
certifi==2018.4.16
chardet==3.0.4
facebook-sdk==2.0.0
idna==2.6
praw==5.1.0
prawcore==0.12.0
requests==2.18.4
update-checker==0.16
urllib3==1.22
